packages/agent/pyproject.toml
[project]
name = "forex-signal-agent"
version = "0.1.0"
description = "Async Forex market movement notifier using Finnhub, Pandas, and TA-Lib with Telegram alerts"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2",
    "numpy>=1.26",
    "httpx>=0.27.0",
    "pydantic>=2.7",
    "APScheduler>=3.10",
    "python-dateutil>=2.9",
    "pytz>=2024.1",
    "aiosqlite>=0.20.0",
    "PyYAML>=6.0.2",
    "TA-Lib>=0.4.28",
    "yfinance>=0.2.54",
    "pandas-stubs==2.3.2.250926",
    "tenacity>=8.2.0",  # For retry mechanisms
]

[project.scripts]
forex-signal-agent = "forex_signal_agent.main:main"
